refactor: StudentDetailが置かれているdomainパッケージをdataパッケージ配下へ移動
