+++ src/main/java/raisetech/student/management/controller/StudentController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import raisetech.student.management.domain.StudentDetail;
+import raisetech.student.management.data.domain.StudentDetail;
 import raisetech.student.management.validation.CreateGroup;
 import raisetech.student.management.validation.UpdateGroup;
 import raisetech.student.management.service.StudentService;
