+++ src/main/java/raisetech/student/management/service/StudentService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import raisetech.student.management.data.Student;
 import raisetech.student.management.data.StudentCourse;
-import raisetech.student.management.domain.StudentDetail;
+import raisetech.student.management.data.domain.StudentDetail;
 import raisetech.student.management.repository.StudentRepository;
 
 /**
